% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseConnectorJars.R
\docType{package}
\name{DatabaseConnectorJars-package}
\alias{DatabaseConnectorJars}
\alias{DatabaseConnectorJars-package}
\title{DatabaseConnectorJars: JAR Dependencies for the 'DatabaseConnector' Package}
\description{
Provides external JAR dependencies for the 'DatabaseConnector' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/DatabaseConnectorJars}
  \item Report bugs at \url{https://github.com/OHDSI/DatabaseConnectorJars/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Marc Suchard
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
  \item Microsoft Inc. (SQL Server JDBC driver) [copyright holder]
  \item PostgreSQL Global Development Group (PostgreSQL JDBC driver) [copyright holder]
  \item Oracle Inc. (Oracle  JDBC driver) [copyright holder]
  \item Amazon Inc. (RedShift  JDBC driver) [copyright holder]
}

}
\keyword{internal}
